+++ m8-calculator/m8-calculator/app/src/main/java/mooc/vandy/java4android/calculator/ui/ActivityInterface.java
@@ -7,12 +7,12 @@ public interface ActivityInterface {
     /**
      * Get the value of the first user input operand.
      */ 
-    int getValueOne();
+    float getValueOne();
 
     /**
      * Get the value of the second user input operand.
-     */ 
-    int getValueTwo();
+     */
+    float getValueTwo();
 
     /**
      * Get the value of the user input operation.
